refactoring unit test for csa profiles and abort excluded combination constraint handling
