+++ data/crac-creation/crac-creator-csa-profiles/src/main/java/com/farao_community/farao/data/crac_creation/creator/csa_profile/crac_creator/cnec/CsaProfileCnecCreator.java
@@ -99,18 +99,18 @@ public class CsaProfileCnecCreator {
         CnecAdder cnecAdder;
         if (CsaProfileConstants.LimitType.CURRENT.equals(limitType)) {
             cnecAdder = crac.newFlowCnec()
-                .withMonitored(false)
-                .withOptimized(true)
-                .withReliabilityMargin(0);
+                    .withMonitored(false)
+                    .withOptimized(true)
+                    .withReliabilityMargin(0);
 
             if (!this.addCurrentLimit(assessedElementId, (FlowCnecAdder) cnecAdder, isCombinableWithContingency)) {
                 return;
             }
         } else if (CsaProfileConstants.LimitType.VOLTAGE.equals(limitType)) {
             cnecAdder = crac.newVoltageCnec()
-                .withMonitored(true)
-                .withOptimized(false)
-                .withReliabilityMargin(0);
+                    .withMonitored(true)
+                    .withOptimized(false)
+                    .withReliabilityMargin(0);
 
             if (!this.addVoltageLimit(assessedElementId, (VoltageCnecAdder) cnecAdder, isCombinableWithContingency)) {
                 return;
@@ -118,9 +118,9 @@ public class CsaProfileCnecCreator {
         } else if (CsaProfileConstants.LimitType.ANGLE.equals(limitType)) {
 
             cnecAdder = crac.newAngleCnec()
-                .withMonitored(true)
-                .withOptimized(false)
-                .withReliabilityMargin(0);
+                    .withMonitored(true)
+                    .withOptimized(false)
+                    .withReliabilityMargin(0);
 
             if (!this.addAngleLimit(assessedElementId, (AngleCnecAdder) cnecAdder, isCombinableWithContingency)) {
                 return;
@@ -151,22 +151,22 @@ public class CsaProfileCnecCreator {
     private void addCnec(CnecAdder cnecAdder, CsaProfileConstants.LimitType limitType, String contingencyId, String assessedElementId, String cnecName, Instant instant, String rejectedLinksAssessedElementContingency) {
         if (CsaProfileConstants.LimitType.CURRENT.equals(limitType)) {
             ((FlowCnecAdder) cnecAdder).withContingency(contingencyId)
-                .withId(cnecName)
-                .withName(cnecName)
-                .withInstant(instant)
-                .add();
+                    .withId(cnecName)
+                    .withName(cnecName)
+                    .withInstant(instant)
+                    .add();
         } else if (CsaProfileConstants.LimitType.VOLTAGE.equals(limitType)) {
             ((VoltageCnecAdder) cnecAdder).withContingency(contingencyId)
-                .withId(cnecName)
-                .withName(cnecName)
-                .withInstant(instant)
-                .add();
+                    .withId(cnecName)
+                    .withName(cnecName)
+                    .withInstant(instant)
+                    .add();
         } else {
             ((AngleCnecAdder) cnecAdder).withContingency(contingencyId)
-                .withId(cnecName)
-                .withName(cnecName)
-                .withInstant(instant)
-                .add();
+                    .withId(cnecName)
+                    .withName(cnecName)
+                    .withInstant(instant)
+                    .add();
         }
 
         if (rejectedLinksAssessedElementContingency.isEmpty()) {
@@ -281,7 +281,7 @@ public class CsaProfileCnecCreator {
             Contingency contingencyToLink = crac.getContingency(contingencyId);
             if (contingencyToLink == null) {
                 csaProfileCnecCreationContexts.add(CsaProfileCnecCreationContext.notImported(assessedElementId, ImportStatus.INCONSISTENCY_IN_DATA, "the contingency "
-                    + contingencyId + " linked to the assessed element doesn't exist in the CRAC"));
+                        + contingencyId + " linked to the assessed element doesn't exist in the CRAC"));
                 return false;
             } else {
                 combinableContingenciesSet.add(contingencyToLink);
@@ -292,7 +292,7 @@ public class CsaProfileCnecCreator {
             Contingency contingencyToRemove = crac.getContingency(contingencyId);
             if (contingencyToRemove == null) {
                 csaProfileCnecCreationContexts.add(CsaProfileCnecCreationContext.notImported(assessedElementId, ImportStatus.INCONSISTENCY_IN_DATA, "the contingency "
-                    + contingencyId + " excluded from the contingencies linked to the assessed element doesn't exist in the CRAC"));
+                        + contingencyId + " excluded from the contingencies linked to the assessed element doesn't exist in the CRAC"));
                 return false;
             } else {
                 combinableContingenciesSet.remove(contingencyToRemove);
@@ -300,7 +300,7 @@ public class CsaProfileCnecCreator {
             }
         }
         csaProfileCnecCreationContexts.add(CsaProfileCnecCreationContext.notImported(assessedElementId, ImportStatus.INCONSISTENCY_IN_DATA, "AssessedElementWithContingency.combinationConstraintKind = "
-            + combinationConstraintKind + " and AssessedElement.isCombinableWithContingency = " + isCombinableWithContingency + " have inconsistent values"));
+                + combinationConstraintKind + " and AssessedElement.isCombinableWithContingency = " + isCombinableWithContingency + " have inconsistent values"));
         return false;
     }
 
@@ -511,13 +511,13 @@ public class CsaProfileCnecCreator {
         String direction = currentLimit.get(CsaProfileConstants.REQUEST_OPERATIONAL_LIMIT_DIRECTION);
         if (CsaProfileConstants.LimitDirectionKind.ABSOLUTE.toString().equals(direction)) {
             flowCnecAdder.newThreshold().withSide(side)
-                .withUnit(Unit.AMPERE)
-                .withMax(normalValue)
-                .withMin(-normalValue).add();
+                    .withUnit(Unit.AMPERE)
+                    .withMax(normalValue)
+                    .withMin(-normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.HIGH.toString().equals(direction)) {
             flowCnecAdder.newThreshold().withSide(side)
-                .withUnit(Unit.AMPERE)
-                .withMax(normalValue).add();
+                    .withUnit(Unit.AMPERE)
+                    .withMax(normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.LOW.toString().equals(direction)) {
             csaProfileCnecCreationContexts.add(CsaProfileCnecCreationContext.notImported(assessedElementId, ImportStatus.NOT_FOR_RAO, "OperationalLimitType.direction is low"));
             return false;
@@ -579,12 +579,12 @@ public class CsaProfileCnecCreator {
         String direction = voltageLimit.get(CsaProfileConstants.REQUEST_OPERATIONAL_LIMIT_DIRECTION);
         if (CsaProfileConstants.LimitDirectionKind.HIGH.toString().equals(direction)) {
             voltageCnecAdder.newThreshold()
-                .withUnit(Unit.KILOVOLT)
-                .withMax(normalValue).add();
+                    .withUnit(Unit.KILOVOLT)
+                    .withMax(normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.LOW.toString().equals(direction)) {
             voltageCnecAdder.newThreshold()
-                .withUnit(Unit.KILOVOLT)
-                .withMin(normalValue).add();
+                    .withUnit(Unit.KILOVOLT)
+                    .withMin(normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.ABSOLUTE.toString().equals(direction)) {
             csaProfileCnecCreationContexts.add(CsaProfileCnecCreationContext.notImported(assessedElementId, ImportStatus.NOT_YET_HANDLED_BY_FARAO, "Only high and low voltage threshold values are handled for now (OperationalLimitType.direction is absolute)"));
             return false;
@@ -610,20 +610,20 @@ public class CsaProfileCnecCreator {
                 return false;
             }
             angleCnecAdder.newThreshold()
-                .withUnit(Unit.DEGREE)
-                .withMax(normalValue).add();
+                    .withUnit(Unit.DEGREE)
+                    .withMax(normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.LOW.toString().equals(direction)) {
             if (handleMissingIsFlowToRefTerminalForNotAbsoluteDirection(assessedElementId, isFlowToRefTerminalIsNull, CsaProfileConstants.LimitDirectionKind.LOW)) {
                 return false;
             }
             angleCnecAdder.newThreshold()
-                .withUnit(Unit.DEGREE)
-                .withMin(-normalValue).add();
+                    .withUnit(Unit.DEGREE)
+                    .withMin(-normalValue).add();
         } else if (CsaProfileConstants.LimitDirectionKind.ABSOLUTE.toString().equals(direction)) {
             angleCnecAdder.newThreshold()
-                .withUnit(Unit.DEGREE)
-                .withMin(-normalValue)
-                .withMax(normalValue).add();
+                    .withUnit(Unit.DEGREE)
+                    .withMin(-normalValue)
+                    .withMax(normalValue).add();
         }
         return true;
     }
